Add parser for uniform or preferred titles
This adds a PreferredTitleParser class for parsing out the various
kinds of preferred titles in fields 130, 240, 243, 700, 710, 711, 730,
740, 800, 810, 811, and 830. (Note the change in terminology from AACR2
to RDA, from "uniform" to "preferred" titles.)

Titles are parsed out into their consistuent hierarchical parts,
depending on both sequence of subfields and ISBD punctuation. Separate
parts are generated with the idea that they'll become separate facets,
so identifying parts that would group things under useful facet
headings (particularly where collective titles are concerned) is key.
Generally, periods at the end of subfields denote the end of a part,
while commas denote the continuation of that part. The exception is
with collective titles, particularly for music forms, where we try to
keep the collective title itself (plus information about music
performance medium) isolated to a single part so that one collective
title facet can be generated later, in combination with the author
name. I.e.: `$aSonatas,$mpiano,$nop. 10` generates two parts:
`Sonatas, piano` and `Op. 10`. The display and faceting would be:
`Sonatas, piano [Author] > Op. 10`. Further, $k always appears alone in
a part by itself.

Collective titles are identified, including whether the collective
title is a music form. These facts will have implications for how work
titles stored in facet/display values are reconstructed, later.

Information related to the expression or version of the work is
separated out from the work title, and an `expression_parts` list is
used to store those values, in the order in which they occur in the
field. This includes $f (dates associated with the work), $l
(language), $o (arranged), $s (version). It also includes anything
appearing after the first instance of one of these subfields; it's rare
but possible to have e.g. a $n or $p following a $s denoting a part of
a specific version. For now we're considering that part of the
expression or version. The intention is that all

Series titles fields may have $v (volume) or $x (issn), and those are
identified if found.

Information that leads to generating small slugs of text for display,
$3 (materials specified) and $i (relationship), are also identified if
found. For $i, any WEMI entity name in parentheses is removed: (work),
(expression), etc.