Determine title type via PreferredTitleParser
The PreferredTitleParser now determines whether a title is a series
title (8XX), an analytic (7XX with ind2 == 2 OR "Container of"
designation), a related title (other 7XXs), or a main title (130, 240,
or 243). If an analytic title has a $i with `Container of`, it is not
included in `display_constants` because it's redundant.