Add capability for fuzzy truncation to str parsers
One aspect of dealing with titles (particularly 245s) is truncating
extremely long titles when needed. This adds a str parser object that
allows you to truncate a string based on a series of patterns that
fall within a certain character range. E.g., to truncate to the nearest
colon, or period, or space within characters 80 to 150.