Add ability to match names from SOR with headings
In order to create full Included Works entries for transcribed titles
(i.e. from 245s) where there are multiple transcribed titles, we have
to be able to match up controlled author headings (i.e. from a 7XX
field) with the appropriate title. SORs are a way to do this. This
commit adds a couple of parsers that parse proper names from an SOR and
help us match one of those names with a heading string.