Fix transcribed titles w/ a non-SOR forward slash
Trying to accommodate edge-cases where the entire title statement was
(for instance) shoved into the 245$a, we were doing a bit too much
parsing based soley on ISBD punctuation, which screwed up patterns
where " / " occurs within the title proper but NOT as the ISBD SOR
indicator. This simplifies things and assumes everything in $a is
just part of the main title and the $b will not have any SORs in it.