Improve music form coll titles and "Selections"
Two improvements rolled into this commit.

1. Music form terms are now divided into singular and corresponding
plural forms. Music collective titles are identified as such only if
they use a plural form of one of the authorized terms. That way, e.g.,
Ravel's Bolero is not considered a collective title.

2. For NON-collective titles that use "Selections" -- such as, "Also
sprach Zarathustra (Selections)" -- the shortened author name used in
the display value was being appended after "(Selections)." For
consistency, it needs to be appended to the title before "(Selections)"
-- i.e., "Also sprach Zarathustra [by Strauss, R.] (Selections)".