Fix handling of non-filing characters in titles
When generating title keys (for sorting and for title/series facet
values), the number of non-filing characters indicated in the title
field causes the title key to omit that many characters from the
beginning of the key string, allowing e.g. "The title" to file under
"title" (assuming the number of nf chars is 4). This change makes two
fixes to this functionality.

1. 245 titles, when there are multiple titles parsed out of one 245,
the num of nf chars can only apply to the first, but previously we were
erroneously applying it to all titles, leading to incorrect keys.

2. Incorrect values entered in the catalog record leads to some weird
behavior. This does a simple reality check to ensure that cutting off
N nf chars from the title string does so at a word break (i.e. a non-
alphanumeric character) and not in the middle of a word.