Fix strange ISBD-related edge cases (titles)
This is an attempt to fix a few odd edge cases when parsing titles
(mostly 245s) using ISBD punctuation, particularly when periods are
in quotation marks, or when other punctuation appears after an ISBD
mark but before the next part of the title (or the end of the title).

Specifically: `Title / by Name "Nickname."` and `Title / by Name. ...`

In both cases, strictly parsing these would generate titles `"` and
`...` respectively, because these appear after the ISBD period. Now
an attempt is made to find these kinds of instances and switch the
offending punctuation with the ISBD mark so that the ISBD mark IS the
end of the title -- e.g., "Nickname." ==> "Nickname".