Fix shortening author first names w/unicode chars
The `shorten_name` function tries generating initials from personal
names (when there is a surname and forename), but in order to make it
deal universally with names that may be hyphenated or already have
initials, it split on non-word characters (regex `\W`). This ended up
splitting on unicode characters inside names so e.g. `Frédéric`
became `F.d.r` instead of `F.` This changes the regex so it explicitly
splits on space or relevant punctuation, to avoid this problem.