Fix title parsing errors caused by typos
After running our entire catalog through the new indexing processes,
we ran into a few (surprisingly few) cases where typos in the MARC
record, such as subfields containing only spaces or punctuation, led to
uncaught parsing errors. This adds a couple of checks to catch those,
ignore that "part" of the title, and keep going rather than throwing
an error that causes an entire chunk of records not to be loaded.