Nail down version of openjdk-8-jre in Docker build
We started seeing calls to Solrmarc hanging during tests after updating
Docker images this past week (and in Travis-CI). I traced the problem
to the last openjdk-8 update, which bumped part of the Java version to
265, and something used by Solrmarc considers versions over 255 to be
illegal, causing it to report the error and then hang when you run it.
In the Dockerfile I tried pinning the version to `8u252-b09-1~deb9u1`
which seems to have fixed the problem. (Note that we're actively
working on removing the Solrmarc dependency entirely.)