Ensure item record updates trigger bib reindexing
This fixes a long-standing (and, now that our Blacklight instance is
live, important) issue where, because reindexing is normally triggered
based on bib records' last-updated date, updating an item record
without touching the parent bib did NOT trigger reindexing for that
record. However, the Solr docs in our Blacklight index contain a
mixture of bib and item information, so changes to items SHOULD trigger
reindexing, just as changes to bibs do.

To fix this, I've added an option to the custom model manager for the
main III record types (i.e. ones that join with the `record_metadata`
table) for the `updated_date_range` custom filter. You can now
(optionally) pass a list of paths from the main record type to attached
records for which you also want to filter based on the last-updated
date range. E.g., so you can create a queryset with bibs that were last
updated within a certain time period OR have an attached item updated
within that time period. And, the blacklight-alpha-solrmarc exporters
now use this feature for exactly this purpose.