Change `filter` call to keep prefetch valid
While testing changes to underlying tasks I noticed exports were
generating thousands of queries, mostly against the `control_field`
table. This didn't make sense since that table is included in our
prefetching. I traced the problem to a call to the `filter` method of
the control_field object manager; I changed that to use a list
comprehension filter instead and fixed the problem.