Make `location` export filter more useful
The `location` export filter previously allowed you to export item
records from Sierra that matched a particular location code. This
improves that feature in two ways. One, a comma-delimited list of
location codes may now be provided, and records matching any of the
provided locations will be exported. Two, the filter now works with
bib-centric as well as item-centric exporters. When using with a bib-
centric exporter, it selects (unique) bibs that have an item at one of
the provided locations. (This should be helpful for managing things
like changes to requesting, as these changes generally apply to sets of
records at specific item locations.)

Also in this commit is an update to how `distinct` works with the
custom filters; rather than forcing the exporters' `get_records`
method to specify `distinct`, certain specific scenarios that may
result in duplicates are detected at the manager/filter level, and
`distinct` is called on the queryset at that time.