Finish custom filter updates for last-updated date
With the updates to the `updated_date_range` custom manager filter, I
completely neglected to check to see what updates to make to the
`last_export` filter, which is what we use most often in production for
regular nightly updates. I'd been thinking that we just used the same
code as `updated_date_range` but realized (when last night's scheduled
update failed to pick up items that had been updated yesterday) this
was not at all the case.

So, this should fix this once and for all. Now there's a generic method
for creating a filter based on a date range, and both the
`updated_date_range` and `last_export` methods use it to generate the
DB query. I also added corresponding tests to test `last_export`
separately.