Improve generating links and thumbs from 962s
Media Manager (in Sierra) files/URLs that are linked to bib records and
represented via 962s in the MARC are slightly tough to deal with,
because they're used for linking to cover images which become
thumbnails in Discover, other full-text or non-full-text URLs (like
856s), and e-reserves. This tweaks the logic for determining what a 962
is and how it should function. Now ANY images are treated as covers
(whereas previously only images in a specific location were). And now
the URLs for e-reserves are generated and stored in the index.