Add Solr field to support pub date range limit
At some point in the not-too-distant future we will be converting the
publication decade and year facets to use the Blacklight range-limit
plugin like nearly everyone else; users do seem to prefer the latter
over the former. To support that, we need a single field with sortable
(i.e. uniformly 4-digit) years. This adds that, along with a copyField
command to copy from the existing publication_year_facet, so that we
can try that.