Make misc structural changes to s2m processes
This makes a few structural changes to some of the more internal
classes/methods in the sierra2marc processes to help facilitate some
of the subject parsing that requires using shared code (such as for
extracting names and titles). It also consolidates handling of the
`materials specified` control subfield (3) so that it isn't necessary
to repeat it for every single parser.

Incidentally also cleaned up whitespace issues.