Add new feature for personal name variations
This adds a new class plus supporting code and tests for generating
variations or permutations of personal names for searching. The goal is
to improve both matching of names (since there are often several valid
ways to enter any given name) and relevance. The class can also be
easily extended as needed if we want to change or update how those
permutations are generated.

See the (new) PersonalNamePermutator class for more information.