Improve author/contributor name searching
This applies much of what we've learned testing searching names as
subjects to regular author/contributor names (personal names,
specifically). The new `PersonalNamePermutator` class is used to
generate search variations. I've also updated the `author_search` and
`contributors_search` fields in Solr to use the `heading_term_text`
field type, which keeps commas as tokens in the index to use to better
match inverted name forms. Finally, author name permutations appear in
both `author_search` and `contributors_search`, to offset the effects
of IDF for names that appear frequently as authors and infrequently as
contributors (the rareness of the name in `contributors_search` leads
to a high IDF value; due to that IDF boost, results where the matching
name appears as a contributor but NOT an author are boosted above
results where the name appears as the main author, which is
counterintuitive).