Fix issue with control subfields displaying
This is a fix for Discover issue 194. Somehow I completely neglected to
exclude control subfield tags (w01256789) from the `PersonalNameParser`
parsing code. This adds them to the list of subfield tags to exclude,
along with tests to ensure this happens for the handful of control
subfields we have in our data.