Improve display of parallel titles from 245
Parallel titles in the MARC 245 -- i.e., versions of the main title
translated into different languages -- are set off by ' = ' in the
catalog record. E.g.: 'Original title = Translation 1 = Translation 2.'

Because of the oddness of this notation, I've wanted to change how we
represent parallel titles in our indexing. Originally I opted to strip
parallel titles from the main title display and just put them in the
alternate titles field. However, sometimes a parallel title contains
the English translation for something like a foreign film, where the
title is unrecognizable without it.

245 notation can get extremely complex when it comes to parallel
titles, particularly when you factor in parallel statements of
responsibility and multiple titles in one 245. This was the main reason
I avoided the issue, previously.

This changeset adds the following.

  - The `TranscribedTitleParser` class `parse` method return value has
    been restructured. Parallel titles are no longer returned as a flat
    list, parallel to the `transcribed` element. Now each title
    structure within the `transcribed` list has its own `parallel`
    element that lists all parallel titles associated with that title.
    This allows us to match any parallel titles with the original title
    that they're translations of, even if there are multiple original
    titles and multiple sets of parallel titles.

  - The `main_title_display` field now renders parallel titles. The
    example given above is rendered: 'Original title [translated:
    Translation 1; Translation 2]'

  - Parallel statements of responsibility are rendered as part of the
    `responsibility_display` field similarly to the rendering for
    the main title.

  - Parallel titles are now included in the `main_title_search` field.

  - For 245s that generate added entries (i.e. Included Works), any
    parallel titles are included in the display value similarly to how
    they're included in the display value for the main title. They also
    create added Title/Series facet values for each individual parallel
    title. They are not included as part of the facet value for the
    original title. (The earlier example would generate three separate
    title/series facet values.)

  - They are NOT included in the relevant title sort values.