Add `combined` name org/event headings
Here's another update to how names are being parsed to make sure they
are compatible with names-as-subjects. This addresses the edge case
where an X10 or X11 field is parsed into an organization *and* an event
name. We want these listed separately for some fields (contributors and
meetings), but for subjects, we want the entire combined org + event to
appear in the display and full-subject facet as a more traditional-
style heading. So, now, when an OrgEventParser is used, if it parses an
organization and an event from the same name heading, it adds a third
result, type `combined`, which will more closely resemble the original
heading.

Also, whenever names *are* parsed, there is now an explicit function
for selecting the best version of the name to use IF there is more than
one. In some cases we prefer the organization, but for subjects we
prefer the combined version.