Update how pref titles are compiled, for subjects
This is yet more preparation to ensure titles will work as intended for
using as subjects. Because of the full-subject-heading facet field and
the fact that the full LCSH heading for name-titles needs to be
preserved, I had to make sure to separate compiling the parts of the
title from rendering facet values and json values. Also, because facet
values for the full-subject-heading facet will have the author name
prepended, I had to move generating the facet key values for title
facets to a higher level method.