Change how `era` terms are handled
This is a small change to how 'era'-type subject terms are mapped. If
era term is found (such as `18th Century`), then the results of the
initial mapping are assumed to be topics, and then provided era term
is added to the return value as an era. This is because the mapping
does not (I believe) include any actual era terms as mapped terms, they
are all topics or genres.