Move `language` field implementation to python
This is the last field that was implemented using purely Solrmarc, now
added to our python pipeline. It improves upon the Solrmarc
implementation in a few ways:

- Languages are collected from the 008[35-37], 041 fields, and any
language fields from uniform/analytic titles (130, 7XXs). The
`language_facet` field is built from a deduplicated list of all
languages in these fields.

- If the record does not supply any 546 (language notes) fields, notes
are generated automatically based on whatever information there is. The
008 language and any languages from analytic titles are considered
"main" languages (equivalent to 041$a or $d). Languages from the 041
are categorized appropriately. One note is generated for each category
of language that has valid language data. Lists of languages are of
course deduplicated on a per-category basis.