Update solrconfig again and add search fields
A few things rolled into this update.

1. Revisions to qf/pf values in solrconfig. The previous revisions work
well, but with the lower boost values I notice that certain records are
getting more of a boost from IDF than we want. This bumps those values
back up a little bit to see if that helps counteract it.

2. Disabled spellcheck in Solr since we aren't using it in Blacklight.

3. Combed through default qf/pf values again and fixed a few things,
such as fieldnames that needed to be updated to match the fields in the
current schema.

4. Added a few new search fields to the schema that simply copy data
from existing (display/facet) fields, to ensure all the information we
need is actually searchable. While preparing to get rid of Solrmarc I
disabled the directive that loaded ALL record data into the catch-all
"text" field, so we need(ed) some new search fields to pick up the
slack.