Tweak search fields in Solr schema and config
Another tweak to some of the search fields. To help control for IDF
variances with titles, I created a combined `all_title_variants_search`
field that includes the main title, included works titles, variant
titles, and series titles. Other than main title I've been giving
these fields nearly the same boost value in `qf` and `pf`, so it makes
sense to combine them from that perspective. Including main title helps
control for having a term or phrase that appears more commonly in the
main title but less frequently as a variant title that then
disproportionately boosts the match in the variant field.

I was seeing this happening with the search term "star wars," both
in the variant title search field and the series title search field.
Hopefully this will help control that issue. The game "Knights of the
Old Republic" was getting a good IDF boost from being one of the few
records with "star wars" in the series title field.