Fix errors in name permutator from bad $q data
Running the updated processes against our full collection illuminated
an additional problem with the name permutator that happens when $q
(fuller form of name) contains "extra" data. E.g., normally it follows
this pattern:
    $aLastname, F.$q(Firstname),$d1960-

Patterns causing problems included:
    $aLastname,$qF. (Firstname)
    $aLastname, F.$q(Firstname), 1960-

I modified the method on `PersonalNameParser` that handles $q and,
rather than simply trying to strip outer parentheses, now it ignores
(and discards) anything outside the parentheses, so both of the above
problem patterns would consider 'Firstname' to be the fuller form of
name. In the first example 'F.' would be stripped, and in the second
', 1960-' would be stripped.

Note that the heading that the `PersonalNameParser` generates still
contains the original data, so this shouldn't affect how names display
at all, just things like the name permutator that generates variations
for searching.