Fix errors from MARC fields missing initial |a tag
Another issue that occurred during the last full reindexing brought to
light that MARC varfield data coming from the Sierra database might
(rarely) be missing the initial `|a` tag from the field content. In
Sierra, the staff display strips this subfield tag and leaves it as
implied, but it should always be present in the underlying data, so
this is some kind of weird Sierra data error.

I've added a step to the process that generates the pymarc record from
the Sierra DB data to check for this. If the field content does not
begin with `|`, then a `|a` tag is added.

This corrects a previously unknown issue with e.g. record b1549537;
the issue affected the main title (245) field, and so accessing the
record in production Discover raises an error. I've reindexed the
record on Alpha and it is now working there.