Make recentness component of `record_boost` fixed
Following the new year I realized that making the "recentness"
component of the `record_boost` integer calculation relative to the
current year was a mistake. A record published in 2020 indexed in 2020
would get a score of 500; a record published in 2020 but indexed in
2021 would get a score of 499. Thus, to work properly, the entire
collection would have to be reindexed following each new year.

Instead, making the score relative to a fixed year (i.e. 2020 always
=> 500) means that the scores won't need to be updated each year. The
upper limit will increase by 1 each year, but this shouldn't be a
problem.