Continue removing SolrMarc dependency
This is the next step in removing the SolrMarc dependency. The goal is
to disable the `BibsToSolr` exporter insofar as it must use SolrMarc to
load data into Solr. This means we will no longer be able to update the
`bibs` API resource. However, we no longer use the `bibs` API resource
for any client applications, so this should not negatively impact us.

The skeleton of the `BibsToSolr` exporter is still needed to hook the
`ItemsToSolr` exporter onto for production loads, in order to ensure
items are updated when the bib records are updated.

In the future we may update the `bibs` resource to use the same index
as Discover, but for now that isn't necessary.