Update utils for importing MARC strings; tests
This is one step toward simplifying how MARC fields are created and
used for tests and testing. The sierra2marc.MarcUtils methods for
converting a string to a set of parameters to be used for generating a
MARC field object have been consolidated and improved:
- A new method, `fieldstring_to_field`, has been added. Instead of
  converting to parameters, it simply returns a SierraMarcField object
  directly.
- The new method now handles MarcEdit-style strings (such as what the
  pymarc Field objects output when cast to strings).
- It now handles control fields, including control field strings that
  specify a character position range, such as `008/18-21`.
- Interpretation of spacing surrounding indicator values has been made
  more deterministic.
- A "printable" version of a field is not longer returned, since that
  can be had simply by printing the SierraMarcField object.
- The `compile_marc_display_field` method has been removed, since
  (again) simply printing the field object works well.

Tests that relied on these methods have been updated. Similarly, test
fixtures that used parameter tuples as an intermediary format to go to
and from field objects have been updated so that field objects are
created (and/or used) directly.

Ultimately, these steps were needed in order to add steps to handle 001
control fields more easily.