Incorporate 001 into control number parsing
This adds the 001 (and 003) into the ToDiscoverPipeline
`get_control_number_info` method, plus tests. The 001 is either an OCLC
or a vendor control number so should be treated like an additional 035
field. However, any 035s that have valid OCLC numbers take precedence,
due to the way we manage holdings in OCLC. For a record with a valid
OCLC number in the 035, an 001 that contains a different OCLC number is
treated as an "invalid" number -- i.e., it's treated as though it were
in an 035$z field.

Duplicate numbers are (of course) not added.