Add ability to parse names without subfield coding
This makes several minor changes to how personal names are parsed and
adds a couple of functions that allow us to (sort of) parse names where
we just have a name heading string and not much else. The purpose of
this is to let us handle names we find in linking fields (76X-78X),
where we only have the heading string and no subfield coding or MARC
indicator values.

The primary use case is to let us create decent shortened author names
that we'll display alongside titles, so we can keep a somewhat
consistent display with other 7XX fields.

`blacklight.parsers.person_title` takes a string that's presumed to be
a personal title and parses a `prefix`, `particle`, and `full_title`,
using the data in `settings.MARCDATA` to make these determinations.
This addresses the use case for generating name permutations for
searching and lets us make a more informed determination whether name
parts such as in "Charles, Prince of Wales" are a forename or title.

`blacklight.sierra2marc.parse_name_string` does the work of taking an
undifferentiated name heading string and parsing it, returning a return
value as though one of the actual `SequentialMarcFieldParser` classes
(for personal names or orgs/events) had parsed it. It uses a bunch of
regular expression heuristics for determining whether something appears
to be a personal or an organizational name, and then parsing it as best
it can. It's not perfect, but will work for generating shortened author
names for linking fields.