Make bundle tests DRYer
I've been wanting to change this for a while -- this adds a new test
fixture for asserting that a bundle or part of a bundle (i.e. output
from running `ToDiscoverPipeline.do`) matches expected values or sets
of values. The similar (existing) fixture for asserting JSON strings
match expected sets of values has been updated so that the two work
similarly. All tests have been updated accordingly.

In addition, while working on this, I caught a minor issue in the
`ToDiscoverPipeline.get_pub_info` method, which I fixed so that tests
would pass.