Fix control number parsing, source and number
This fixes an issue parsing control numbers like:

`(Abc-Def)12345678`

Where the MARC organization code representing the source of the control
number (in parentheses) is not parsed out correctly. I found that the
pattern used for matching this was too restrictive: it only looked for
"word" characters (\w), thus excluding hyphens, which are valid.

There is no reason for this pattern-matching to be any more restrictive
than simply looking for anything in parentheses. I have updated the
behavior accordingly.

Further, because control numbers formatted this way can appear in
fields other than the 035 (such as linking fields), I pulled that out
into a class method on the StandardControlNumberParser and updated the
processes that parse the 035s and the ones that process the linking
fields so they both use that method.