Fix erroneous years from 362s
We noticed some erroneous 3-digit publication years coming through the
publication facets. Part of the problem is that we're currently trying
to pull years from 362 fields, which may contain 3- or 4-digit volume
numbers that get mistaken for years. After looking through the examples
I don't really see a reason to continue pulling years from 362s, as the
legitimate years are generally also present in an 008, 046, 260, or 264
field.