Remove relators from linking field author searches
When indexing names for author_search and contributors_search fields,
we don't index relators as part of the heading -- but some linking
fields include them as part of the author name in $a, which leads to
failed searches when using the full $a in the search criteria.

To correct the problem, this changeset adds an extra step to
`parse_name_string` that chops off the relators (presumably a comma-
delimited, all lowercase list of words) from the end of the author
string and converts them to a list, mimicking the output of
`PersonalNameParser`. Now we can just use the author heading minus
relators as the search term from the linking field $a.