Add MARC 866 display field to discover Solr schema
Our catalog uses 866 fields to store data (in the bib record) about
serial holdings. (Separate holdings records are also used, but the 866
is the least complicated source for getting this information.)

Colloquially we've always called this the "Library Has" field. So this
change adds a `library_has_display` field to Solr. Each individual 866
field on the record is added ($az only).

It's on our future to-do list to begin exploring actual holdings
records to incorporate that information into Discover, but this is a
start.