Add 383 and 384 (music information)
This adds the 383, which contains the numeric designation for musical
works, and 384, which contains the key. Often this info is represented
in a title field -- and often the 383 and 384 is duplicative of that
title info -- so, for now, we are dumping this info into
`variant_titles_search` IF it isn't already present in the main title
or an analytic title.