Ensure mat specified + display label is consistent
Primarily this updates the GenericDisplayFieldParser and related code
so that display labels are formatted consistently, particularly when
there is a $3 materials specified AND a display label/constant. It
should be formatted: `(mat specified) Label: Field content`.

Updating this allowed me to further simplify some of the code that
generates 5xx notes fields values.