Compile MARC leader via export processes
Non-system-generated leader values from the Sierra record were not
being compiled and added to the MARC record as part of the process that
reconsitutes MARC records (via pymarc) from the Sierra database. This
adds it, at the expense of adding an additional DB relationship to the
query for that exporter.