Use Django 1.9 field relationship descriptors ...
... for inpecting models via the `relationtrees` module.

Django 1.9 changes the `django.db.models.fields.related` API to
`django.db.models.fields.related_descriptors`, which appears to hold
through current Django (3.2). This changeset updates the code in
`relationtrees.Relation` that inspects and tries to characterize the
Django field relationship to use the new API. It also adds a handful of
additional test cases to make sure it's tested as thoroughly as
possible.