Add basic integration tests for Celery tasks
This changeset adds a new test file and adds/updates some test fixtures.
The goal is to have integration tests for Celery tasks that are as close
as possible to the real thing, to help troubleshoot Celery problems that
won't arise during the other (unit) tests. It tests each exporter,
running a full_export, triggered via the `export_dispatch` task, running
on a celery worker that is identical to the production worker.