Fix item sorting for null/duplicate display order
Generally we want item records to sort in the display order set in
Sierra, reflected in the DB via the bib/item link `items_display_order`
integer value. However, sometimes there are null or duplicate values in
the data. This fixes sorting item records to account for this: 1) null
values now default to 0, which causes these items to display first; 2)
the item record number is now a secondary sort, so groups of items with
the same display_order value sort internally by record number. All else
being equal, record number order is a good default order, because it
puts things in the order in which they were created.