Fix base/test/test_managers.py failures
Past versions of Django didn't seem to care that I was being careless
with how I was creating some of the temporary fixtures in this test
file. Specifically, creating blank/empty location instances for items
lacking locations. In Django 3.2, this started to cause an error. Now,
if an item does not have a valid location code, then the value defaults
to None (i.e. null) instead of some unknown blank-ish location instance.
This fixes the error; these tests pass now.