Fix datetime conversion for pysolr 3.9.0
Pysolr 3.9.0 switches to the JSON API (from XML) for adding records to
Solr, but the JSON encoder doesn't handle converting datetime objects to
strings, where the XML method did/does. We could just switch back to the
XML method, but the JSON method is supposed to be a lot faster.

The only place we're seeing an error from this is when generating test
fixtures and loading them into Solr, so this adds a (kind of hacky) fix,
which is to convert datetime objects in test records just before saving
them to Solr via the fixture factory.