Migrate Solr tooling to upgrade to 8.10
This gets Solr upgraded to the latest version (8.10) and moves us to the
official Docker image. There's still tons to do to migrate our
cores/collections, but this takes care of the application itself.

Two big things in this changeset.

First, a big refactoring of the docker-compose service configuration
and Docker build to allow us a central location to define services and
paths: the `docker-compose.env` file. All of the service-specific
definition we had in `init-dockerdata.sh` has been moved into
`docker-compose.env`, which both `init-dockerdata.sh` and
`docker-compose.sh` source. Now, we should be able to maintain all of
the path information in the env file. (We'll still have to update
`docker-compose.yml` when we add or change services.)

Second, this completes a big step in the Solr upgrade process. Solr
services have now been switched to the `solr:8.10.0` Docker image, and
the setup for these services is now working — except of course that the
schema and config files need to be updated to work with Solr 8. Now,
instead of relying on the install of Solr in `/solr/instances`, we can
just have a handful of configuration files (`/solrconf`). For Docker
Solr, I've set up symlinks to these files (in `docker_data`) so that we
can still update the configuration in the repository directly without
having to reinitialize data. In production, we'll do something similar —
symlink to the configuration files from the cores in the Solr home
directory.

Note that any tests that require Solr fail, because the Solr setup is
not finished. (Tests that don't require Solr pass.)