Update Discover solr conf files for Solr 8.10
This updates the `discover-01` and `discover-02` solrconfig.xml and
schema.xml files for Solr 8.10.

At this point I'm setting things up using recommended settings as much
as possible, e.g. docValues for all sortable/facetable fields. Later
we'll need to do some performance testing to see how well this performs
compared to the old configuration.

Notes about a couple of things:

- I've removed maxWarmingSearchers from solrconfig.xml. According to
https://solr.apache.org/guide/7_0/major-changes-in-solr-7.html, commits
now block so that there is no longer a chance for overlapping commits,
and the maxWarmingSearchers parameter is no longer needed. The
recommendation is to remove it.

- I did a little bit of research/testing/analysis on the (newer)
WordDelimiterGraphFilterFactory and made a few tweaks to the
implementation. I wasn't sure exactly where the
FlattenGraphFilterFactory needed to go and whether its inclusion made
any difference for the query analyzer.

Following these changes, the cores now load. (We may need further
updates as we work through issues with tests and then actual record
loads, etc.)