Refactor: move `blacklight` code to `export` app
One of my goals with the Solr upgrade is to remove the `bibdata` core
and update the functionality that used it (like certain API resources)
so that it uses the `discover` cores instead. As a first step on that
path, we have to merge the entire contents of the `blacklight` app into
the main app, mainly `export`. (This commit.) Later we'll also have to
update the applicable API resources so they use the `discover` cores
instead of the `bibdata` core. (In subsequent commits.)

Notes about the refactor:

- Because `blacklight.sierra2marc` was so large, I'm splitting this up
into multiple files/parts under a new `export.marcparse` module/package.
Ideally we'll eventually be able to package that up (or parts of it) by
itself to provide some standalone MARC parsing tools.

- I tried to update terminology as much as possible to remove the term
`blacklight`.

- Made a few additional fixes to get the task integration tests working:

    - Switch all BibsToDiscover exporters to BibsToSolr.
    - Update hardcoded `discover-02` core to `settings.BL_CONN_NAME` to
      pick up and use whatever the current BL/Discover core is.
    - Fixed error(s) with Haystack schema related to the Solr upgrade,
      added `docValues="true"` declarations for all potentially sortable
      fields.
    - Fixed odd int/float/string conversion problem with the
      `record_boost` field; I think this might hint at some of the other
      int/float/string issues we're having following upgrading to Py3.